Add precommit hook and fix resultant issues

+++ DICOM_solver/DVH/dicom_bundle.py
@@ -1,37 +1,36 @@
+import logging
 import os
 
 from dicompylercore.dicomparser import DicomParser
-import logging
 
 
 class DicomBundle:
-
-    def __init__(self, patient_id, rt_plan: str, rt_struct: str, rt_dose: [],
-                 rt_ct: str, read=True):
+    def __init__(self, patient_id, rt_plan: str, rt_struct: str, rt_dose: list, rt_ct: str, read=True):
         self.patient_id = patient_id
         self.rt_plan_path = rt_plan
         self.rt_struct_path: str = rt_struct
-        self.rt_ct_path: str = rt_ct[:rt_ct.rindex("/") + 1] if rt_ct else None
+        self.rt_ct_path: str | None = rt_ct[: rt_ct.rindex("/") + 1] if rt_ct else None
         if read:
             try:
                 self.rt_plan: DicomParser = DicomParser(rt_plan)
                 self.rt_struct: DicomParser = DicomParser(rt_struct)
-                self.rt_dose: [] = [DicomParser(rt) for rt in rt_dose]
-                self.rt_dose_path: [] = rt_dose
+                self.rt_dose: list = [DicomParser(rt) for rt in rt_dose]
+                self.rt_dose_path: list = rt_dose
             except Exception as e:
-                logging.error(f"Error reading DICOM files: {e}")
+                logging.exception(f"Error reading DICOM files: {e}")
                 raise e
         logging.info(f"Ct path is {self.rt_ct_path}")
 
     def __eq__(self, other):
         if not isinstance(other, DicomBundle):
             return False
-        if self.rt_plan_path == other.rt_plan_path and self.rt_ct_path == other.rt_ct_path and \
-                self.rt_struct_path == other.rt_struct_path:
-
+        if (
+            self.rt_plan_path == other.rt_plan_path
+            and self.rt_ct_path == other.rt_ct_path
+            and self.rt_struct_path == other.rt_struct_path
+        ):
             return True
-        else:
-            return False
+        return False
 
     # function to delete all the elemnt using the path of each element
     def rm_data_patient(self):
@@ -44,11 +43,8 @@ class DicomBundle:
             for rt in self.rt_dose_path:
                 logging.info(f"Removing data rt dose {self.rt_dose_path}")
                 os.remove(rt)
-            if self.rt_plan_path is not None:
+            if self.rt_plan_path is not None and self.rt_ct_path is not None:
                 for f in os.listdir(self.rt_ct_path):
                     os.remove(os.path.join(self.rt_ct_path, f))
         except Exception as e:
             logging.warning(f"Error deleting files: {e}")
-
-
-
